m_unicode (m_unicode v0.1.0)

Functions

characters_to_binary(Input)

Converts a possibly unicode list of integers into a binary representing Unicode characters. Returns binary on success or 'error' atom otherwise.

characters_to_binary(Input, InEncoding)

characters_to_list(Input)

Converts a possibly deep list of integers and binaries into a list of integers representing Unicode characters. Returns unicode character list on success or 'error' atom otherwise. The binaries in the input can have characters encoded as one of the following: - ISO Latin-1 (0-255, one character per byte). Here, case parameter InEncoding is to be specified as latin1. - One of the UTF-encodings, which is specified as parameter InEncoding.

characters_to_list(Input, InEncoding)

explode_trim_skip_empty(Search, Subject)

is_ascii(T)

Returns true if the specified unicode list contains only ASCII characters or false otherwise.

is_ascii_alpha(T)

Returns true if the specified unicode list contains only ASCII alphabetic characters or false otherwise.

is_ascii_alpha_lower(T)

Returns true if the specified unicode list contains only ASCII alphabetic lower case characters or false otherwise.

is_ascii_alpha_numeric(T)

Returns true if the specified unicode list contains only ASCII alphabetic and numeric characters or false otherwise.

is_ascii_alpha_numeric_lower(T)

Returns true if the specified unicode list contains only ASCII alphabetic lower case and numeric characters or false otherwise.

is_ascii_alpha_numeric_upper(T)

Returns true if the specified unicode list contains only ASCII alphabetic upper case and numeric characters or false otherwise.

is_ascii_alpha_upper(T)

Returns true if the specified unicode list contains only ASCII alphabetic upper case characters or false otherwise.

is_ascii_numeric(T)

Returns true if the specified unicode list contains only ASCII numeric characters or false otherwise.

is_unicode_alpha(T)

Returns true if the specified unicode list contains only alphabetic letters or false otherwise. All letters in all languages are taken in account not only standard english ASCII letters. This function allow you to check for valid names for example.

is_unicode_name(T)

replace(T, Replace, Subject)

Replace the specified list of characters with the specified character or characters

replace_string(Search, Replace, Subject)

Replace the specified string with the specified character or characters

to_lower(UnicodeList)

to_title(UnicodeList)

to_upper(UnicodeList)

trim(UnicodeList)

Trim unicode list

trim(UnicodeList, _)

Trim unicode list left, right, both or none

trim_left(UnicodeList)

Left trim unicode list

trim_right(UnicodeList)

Right trim unicode list