m_unicode_char (m_unicode v0.1.0)

Functions

is_ascii(UnicodeChar)

Returns true if the specified unicode character is an ASCII character or false otherwise.

is_ascii_alpha(UnicodeChar)

Returns true if the specified unicode character is an ASCII alphabetic character or false otherwise.

is_ascii_alpha_lower(UnicodeChar)

Returns true if the specified unicode character is an ASCII alphabetic lower case character or false otherwise.

is_ascii_alpha_numeric(UnicodeChar)

Returns true if the specified unicode character is an ASCII alpha numeric character or false otherwise.

is_ascii_alpha_numeric_lower(UnicodeChar)

Returns true if the specified unicode character is an ASCII alpha numeric lower character or false otherwise.

is_ascii_alpha_numeric_upper(UnicodeChar)

Returns true if the specified unicode character is an ASCII alpha numeric upper character or false otherwise.

is_ascii_alpha_upper(UnicodeChar)

Returns true if the specified unicode character is an ASCII alphabetic upper case character or false otherwise.

is_ascii_numeric(UnicodeChar)

Returns true if the specified unicode character is an ASCII numeric character or false otherwise.

is_unicode_alpha(UnicodeChar)

Returns true if the specified character is valid alphabetic unicode letter or false otherwise

is_unicode_alpha_numeric(UnicodeChar)

Returns true if the specified unicode character is valid alpha numeric unicode letter or false otherwise

is_unicode_line_break(_)

Returns true if the specified unicode character is a line break unicode character or false otherwise. This method will return true if the character is one of the following: - line feed - carriage return - next line - line separator 8232 - Common General Punctuation Separator space. Line. - paragraph separator 8233 - Common General Punctuation Separator space. Paragraph.

is_unicode_numeric(UnicodeChar)

Returns true if the specified character is valid numeric unicode letter or false otherwise

is_unicode_space(_)

Returns true if the specified unicode character is a spacing unicode character or false otherwise this method will return true if the character is one of the following: - character tabulation - line tabulation - form feed - space Most common (normal ASCII space) - no-break space, identical to 16#0020, but not a point at which a line may be broken. - ogham space mark. Used for interword separation in Ogham text. Normally a vertical line in vertical text or a horizontal line in horizontal text, but may also be a blank space in "stemless" fonts. - en quad 8192 - Punctuation Separator space Width of one en. - em quad 8193 - Common General Punctuation Separator space Also known as "mutton quad". Width of one em. - en space 8194 - Common General Punctuation Separator space Also known as "nut". Width of one en. - em space 8195 - Common General Punctuation Separator space. Also known as "mutton". Width of one em. - three-per-em space 8196 - Common General Punctuation Separator space. Also known as "thick space". One third of an em wide. - four-per-em space 8197 - Common General Punctuation Separator space. Also known as "mid space". One fourth of an em wide. - six-per-em space 8198 - Common General Punctuation Separator space. One sixth of an em wide. In computer typography, sometimes equated to 16#2009. - figure space 8199 - Common General Punctuation Separator space. Figure space. In fonts with monospaced digits, equal to the width of one digit. HTML/XML named entity: numsp; - punctuation space 8200 - Common General Punctuation Separator space. As wide as the narrow punctuation in a font, HTML/XML named entity: puncsp; - thin space 8201 - Common General Punctuation Separator space. One-fifth (sometimes one-sixth) of an em wide. Recommended for use as a thousands separator for measures made with SI units. - hair space 8202 - Common General Punctuation Separator space. Thinner than a thin space. HTML/XML named entity: hairsp; - narrow no-break space 8239 - Common General Punctuation Separator space. Narrow no-break space. Similar in function to 16#00A0 No-Break Space. When used with Mongolian, its width is usually one third of the normal space. - medium mathematical space 8287 - Common General Punctuation Separator space. MMSP. Used in mathematical formulae. - ideographic space 12288 - CJK Symbols Punctuation Separator, used, for example, in tai tou.

is_unicode_space_or_line_break(_)

Returns true if the specified unicode character is a spacing unicode character or a line break unicode character. this method will return true if the character is one of the following: - character tabulation - line feed - line tabulation - form feed - carriage return - space Most common (normal ASCII space) - next line - no-break space, identical to 16#0020, but not a point at which a line may be broken. - ogham space mark. Used for interword separation in Ogham text. Normally a vertical line in vertical text or a horizontal line in horizontal text, but may also be a blank space in "stemless" fonts. - en quad 8192 - Punctuation Separator space Width of one en. - em quad 8193 - Common General Punctuation Separator space Also known as "mutton quad". Width of one em. - en space 8194 - Common General Punctuation Separator space Also known as "nut". Width of one en. - em space 8195 - Common General Punctuation Separator space. Also known as "mutton". Width of one em. - three-per-em space 8196 - Common General Punctuation Separator space. Also known as "thick space". One third of an em wide. - four-per-em space 8197 - Common General Punctuation Separator space. Also known as "mid space". One fourth of an em wide. - six-per-em space 8198 - Common General Punctuation Separator space. One sixth of an em wide. In computer typography, sometimes equated to 16#2009. - figure space 8199 - Common General Punctuation Separator space. Figure space. In fonts with monospaced digits, equal to the width of one digit. HTML/XML named entity: numsp; - punctuation space 8200 - Common General Punctuation Separator space. As wide as the narrow punctuation in a font, HTML/XML named entity: puncsp; - thin space 8201 - Common General Punctuation Separator space. One-fifth (sometimes one-sixth) of an em wide. Recommended for use as a thousands separator for measures made with SI units. - hair space 8202 - Common General Punctuation Separator space. Thinner than a thin space. HTML/XML named entity: hairsp; - line separator 8232 - Common General Punctuation Separator space. Line. - paragraph separator 8233 - Common General Punctuation Separator space. Paragraph. - narrow no-break space 8239 - Common General Punctuation Separator space. Narrow no-break space. Similar in function to 16#00A0 No-Break Space. When used with Mongolian, its width is usually one third of the normal space. - medium mathematical space 8287 - Common General Punctuation Separator space. MMSP. Used in mathematical formulae. - ideographic space 12288 - CJK Symbols Punctuation Separator, used, for example, in tai tou.

to_lower(C)

to_title(C)

to_upper(C)